+++ lib/metronome_platform_interface.dart
@@ -3,6 +3,7 @@ import 'dart:async';
 import 'package:plugin_platform_interface/plugin_platform_interface.dart';
 
 import 'metronome_method_channel.dart';
+import 'tempo_ramp.dart';
 
 abstract class MetronomePlatform extends PlatformInterface {
   /// Constructs a MetronomePlatform.
@@ -27,6 +28,8 @@ abstract class MetronomePlatform extends PlatformInterface {
 
   final StreamController<int> tickController =
       StreamController<int>.broadcast();
+  final StreamController<TempoRampProgress> tempoRampController =
+      StreamController<TempoRampProgress>.broadcast();
 
   Future<void> init(
     String mainPath, {
@@ -78,6 +81,16 @@ abstract class MetronomePlatform extends PlatformInterface {
 
   Future<int?> getBPM() {
     throw UnimplementedError('getBPM() has not been implemented.');
+  }
+
+  Future<void> configureTempoRamp(TempoRampConfig config) {
+    throw UnimplementedError(
+      'configureTempoRamp() has not been implemented.',
+    );
+  }
+
+  Future<void> disableTempoRamp() {
+    throw UnimplementedError('disableTempoRamp() has not been implemented.');
   }
 
   Future<void> setTimeSignature(int timeSignature) {